size_t purge

+++ src/main.c
@@ -9,9 +9,9 @@ int main(void) {
     const char* fp_travel = "./data/dados_distancias.in";
     const char* fp_clubes = "./data/dados_times.in";
 
-    size_t Nclubes = getlines(fp_clubes);
-    size_t Ncities = getlines(fp_cities);
-    size_t Nrodada = getRodadas(Nclubes);
+    int Nclubes = getlines(fp_clubes);
+    int Ncities = getlines(fp_cities);
+    int Nrodada = getRodadas(Nclubes);
     int** tabela = allocTable(Nclubes,Nrodada);
     int** travel = allocTable(Ncities,Ncities);
     cities_t cities[Ncities];
