+++ src/mando.h
@@ -1,4 +1,4 @@
-void mando(size_t n, int** tabela);
-void sorteiaRDN(size_t n, int** tabela, int rodada);
-void espelha(size_t n, int** tabela);
+void mando(int n, int** tabela);
+void sorteiaRDN(int n, int** tabela, int rodada);
+void espelha(int n, int** tabela);
 int randint(int n);
