+++ src/tools.c
@@ -1,6 +1,6 @@
 #include "main.h"
 
-int** allocTable(size_t n, size_t m) {
+int** allocTable(int n, int m) {
     int i; 
     int **table = calloc(n,sizeof(int*));
     if (table == NULL) exit(EXIT_FAILURE);
@@ -9,14 +9,14 @@ int** allocTable(size_t n, size_t m) {
     return table;
 }
 
-void freeMemory(size_t n, int** ptr) {
+void freeMemory(int n, int** ptr) {
     int i;
     for(i=0 ; i<n ; i++)
         free(ptr[i]);
     free(ptr);
 }
 
-void shiftArray(size_t n, int* array) {
+void shiftArray(int n, int* array) {
     int tmp = array[n-1];
     memmove(&array[1], &array[0], (n-1)*sizeof(int));
     array[0] = tmp;
